feat(badge): allow string labels
