+++ src/components/badge/badge.tsx
@@ -1,8 +1,11 @@
-import { Component, Prop } from '@stencil/core';
+import { Component, Prop, h } from '@stencil/core';
 import { abbreviate } from './format';
 
 /**
- * @exampleComponent limel-example-badge
+ * The Badge component can display both `number` and `string` as `label`.
+ *
+ * @exampleComponent limel-example-badge-number
+ * @exampleComponent limel-example-badge-string
  */
 @Component({
     tag: 'limel-badge',
@@ -11,12 +14,19 @@ import { abbreviate } from './format';
 })
 export class Badge {
     /**
-     * Label to display in the badge
+     * Label to display in the badge.
+     * Numeric labels larger than 999 will be rounded and abbreviated.
+     * String labels get truncated if their visual length is longer than
+     * six characters (six `0`s to be exact).
      */
     @Prop({ reflect: true })
-    public label: number;
+    public label: number | string;
 
-    render() {
-        return abbreviate(this.label);
+    public render() {
+        if (typeof this.label === 'number') {
+            return <span>{abbreviate(this.label)}</span>;
+        } else if (typeof this.label === 'string') {
+            return <span>{this.label}</span>;
+        }
     }
 }
