+++ src/components/badge/examples/badge-string.scss
@@ -1,0 +1,21 @@
+:host(limel-example-badge-string) {
+    display: flex;
+    justify-content: space-between;
+}
+
+limel-badge {
+    &:nth-of-type(1) {
+        --badge-background-color: rgb(var(--color-magenta-light));
+        --badge-text-color: rgb(var(--color-yellow-lighter));
+    }
+
+    &:nth-of-type(2) {
+        --badge-background-color: rgb(var(--color-amber-light));
+        --badge-text-color: rgb(var(--color-black));
+    }
+
+    &:nth-of-type(3) {
+        --badge-background-color: rgb(var(--color-violet-default));
+        --badge-text-color: rgb(var(--color-white));
+    }
+}
