--- src/components/badge/examples/badge.scss
@@ -1,19 +1,0 @@
-:host(limel-example-badge) {
-    display: flex;
-    justify-content: space-between;
-}
-
-.badge:nth-of-type(2) {
-    --badge-background-color: rgb(var(--color-red-default));
-    --badge-text-color: rgb(var(--color-yellow-lighter));
-}
-
-.badge:nth-of-type(3) {
-    --badge-background-color: rgb(var(--color-green-default));
-    --badge-text-color: rgb(var(--color-gray-lighter));
-}
-
-.badge:nth-of-type(4) {
-    --badge-background-color: rgb(var(--color-blue-default));
-    --badge-text-color: rgb(var(--color-white));
-}
